Implement InterestManagerCompoundV2
InterestManagerCompoundV2 implements significant gas savings.
The previous version of the InterestManagerCompound contract invested Dai into Compound
on every invest() call. With the current gas prices this increases the cost of each trade
on Ideamarket by $40-$80.

To reduce this cost to the users, this new V2 contract does not invest Dai into Compound
on every call anymore. Instead, it holds the Dai until supplyDaiToCompound() is manually called.

The resulting difference in interest is spread across all IdeaTokens.